Rename file for tests to test_*
The Python2 inittest need the testing files starting with "test_".
Simply renaming the one existing file in test/ so it is matching the
required rule.
