+++ test/__init__.py
@@ -1,4 +1,4 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .client import HieraClientTests
+from .test_client import HieraClientTests
